extract embed button
