+++ client/js/_author/components/bank_navigation/bank_list_item.jsx
@@ -1,10 +1,10 @@
 import React            from 'react';
 import _                from 'lodash';
-import CopyToClipboard  from 'react-copy-to-clipboard';
 import appHistory       from '../../history';
 import Icon             from './bank_icon';
 
 import PublishButton    from './buttons/publish_button';
+import EmbedButton      from './buttons/embed_button';
 
 export function ListItem(props) {
   const { selectItem, bank } = props;
@@ -34,55 +34,6 @@ export function BankFolder(props) {
     </ListItem>
   );
 }
-
-
-  function EmbedButton(props) {
-    const { assessment, baseEmbedUrl, getEmbedCode } = props;
-    const isPublished = assessment.isPublished;
-    const assessOffered = _.get(assessment, 'assessmentOffered[0]');
-
-    if (isPublished) {
-      if (assessOffered) {
-        const embedUrlCode = `${baseEmbedUrl}&bank=${assessOffered.bankId}&assessment_offered_id=${assessOffered.id}#/assessment`;
-
-        return (
-          <div className="au-c-embed-contain">
-            <label className="au-c-input--purple" htmlFor="embedInput">
-              <input
-                id="embedInput"
-                onClick={e => e.stopPropagation()}
-                className="au-c-text-input au-c-text-input--smaller"
-                readOnly
-                type="text"
-                value={`<iframe src="${embedUrlCode}"/>`}
-              />
-            </label>
-            <CopyToClipboard text={`<iframe src="${embedUrlCode}"/>`}>
-              <button
-                className="au-c-btn au-c-btn--square au-c-btn--embed "
-                onClick={e => e.stopPropagation()}
-              >
-                <i className="material-icons">content_paste</i>
-              </button>
-            </CopyToClipboard>
-          </div>
-        );
-      }
-
-      return (
-        <button
-          className="au-c-btn au-c-btn--sm au-c-btn--table"
-          onClick={(e) => {
-            e.stopPropagation();
-            getEmbedCode(assessment);
-          }}
-        >
-          embed code
-        </button>
-      );
-    }
-    return null;
-  }
 
 export function EditButton(props) {
   const isPublished = props.assessment.isPublished;
